+++ lotus/responses.py
@@ -3,6 +3,6 @@ from django.http import HttpResponse
 
 class HttpResponseUnauthorized(HttpResponse):
     """
-    Response to implemente the Unauthorized HTTP status code formerly.
+    Response to implement the Unauthorized HTTP status code formerly.
     """
     status_code = 401