+++ lotus/urls.py
@@ -10,6 +10,7 @@ from .views import (
     PreviewTogglerView, PreviewArticleDetailView,
     TagIndexView, TagDetailView, TagAutocompleteView,
 )
+from .views.category import CategoryTreeView
 
 
 app_name = "lotus"
@@ -26,6 +27,10 @@ urlpatterns = [
     ),
 
     path("categories/", CategoryIndexView.as_view(), name="category-index"),
+
+    # WARNING: Probably temporarily only
+    path("categories/tree/", CategoryTreeView.as_view(), name="category-tree"),
+
     path(
         "categories/<slug:slug>/",
         CategoryDetailView.as_view(),