+++ lotus/utils/jsons.py
@@ -10,6 +10,10 @@ from ..models import Album, Article, Author, Category
 class ExtendedJsonEncoder(json.JSONEncoder):
     """
     Additional opiniated support for more basic object types.
+
+    Usage sample: ::
+
+        json.dumps(..., cls=ExtendedJsonEncoder)
     """
     def default(self, obj):
         if isinstance(obj, bytes):