+++ lotus/views/category.py
@@ -101,3 +101,60 @@ class CategoryDetailView(BaseBreadcrumbMixin, ArticleFilterAbstractView,
 
         # Let the ListView mechanics manage list pagination from given queryset
         return super().get(request, *args, **kwargs)
+
+
+# WARNING: Temporarily during development
+from django.http import HttpResponse
+from django.views.generic import View
+class CategoryTreeView(LotusContextStage, PreviewModeMixin, LanguageMixin, View):
+    """
+    Tree of categories in plain/text format.
+
+    Categories are properly filtered on language since anyway we can't display a tree
+    of all categories with mixed languages.
+    """
+    model = Category
+    lotus_stage = "categories"
+
+    def get_queryset(self):
+        """
+        Build queryset base with language filtering to list categories.
+        """
+        q = self.model.objects.get_for_lang(self.get_language_code())
+
+        # NOTE: Deprecated usage of ``*self.model.COMMON_ORDER_BY`` in order_by since
+        # treebeard is used to add new node sorted on title, so path resolution should
+        # be identical to the previous behavior before treebeard implementation.
+        return q.order_by("path")
+
+    def get_output(self):
+        """
+        Build output items
+        """
+        title = "For language '{}'".format(self.get_language_code())
+
+        output = [
+            title,
+            "=" * len(title),
+            "",
+        ]
+
+        items = Category.get_annotated_list_qs(self.get_queryset())
+        row = "{indent}{name}"
+        for obj, data in items:
+            indent = ("    " * data["level"]) + "└── "
+            output.append(row.format(
+                indent=indent if data["level"] > 0 else "",
+                name=obj.slug,
+            ))
+
+        return output
+
+    def get(self, request):
+        """
+        Return built JSON content from features.
+        """
+        return HttpResponse(
+            "\n".join(self.get_output()),
+            content_type="text/plain; charset=utf-8"
+        )